requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Fiona==1.8.21
fonttools==4.29.1
frozenlist==1.3.0
geoip2==4.6.0
geopandas==0.10.2
idna==3.3
importlib-metadata==4.12.0
ipaddress==1.0.23
kiwisolver==1.3.2
matplotlib==3.5.1
maxminddb==2.2.0
multidict==6.0.2
munch==2.5.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pkg_resources==0.0.0
py4j==0.10.9.2
pycountry==22.3.5
pyparsing==3.0.7
pyproj==3.2.1
pyspark==3.2.0
pysubnettree==0.35
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
scipy==1.7.3
seaborn==0.11.2
Shapely==1.8.2
six==1.16.0
typing_extensions==4.3.0
UpSetPlot==0.6.1
urllib3==1.26.10
yarl==1.7.2
zipp==3.8.1
